backend/requirements.txt
# Core FastAPI dependencies (updated for Python 3.13 compatibility)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-multipart>=0.0.12

# Data science libraries (compatible versions)
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.2

# Google Gemini for AI-powered insights
google-generativeai>=0.8.0

# Real market data
yfinance>=0.2.40
httpx>=0.27.0

# ML/AI libraries (optional - only if needed)
# Uncomment these if you have GPU and want to use FinGPT models
# torch>=2.1.0
# transformers>=4.35.0
# peft>=0.7.1
# accelerate>=0.25.0
# sentencepiece>=0.1.99
# protobuf>=4.25.0
